requirements.txt
streamlit>=1.36.0
anthropic>=0.40.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0
Pillow>=9.0.0
# Google Sheets integration
gspread>=6.0.0
google-auth>=2.29.0
# Webhook (Railway service)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
